change exit_code return type from i32 to u8
Change FormatResult::exit_code() return type from i32 to u8, as the
function only returns values 0, 1, or 2, which fit within u8 range.

This eliminates the need for type casting in main.rs and makes the
API more accurate.

Changes:
- Update exit_code() return type: i32 -> u8
- Remove unnecessary type cast in main.rs

+++ src/main.rs
@@ -22,7 +22,7 @@ fn main() -> ExitCode {
     };
 
     match result {
-        Ok(result) => ExitCode::from(result.exit_code() as u8),
+        Ok(result) => ExitCode::from(result.exit_code()),
         Err(_) => ExitCode::from(2),
     }
 }
